+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### Version 0.4
+Le 05/12/2015
+
+- kinetik : différencie les chemins
+- remplace les virgules par des points (erreur float)
+- integre la première colonne (del5q)
+
 ### Version 0.3
 Le 03/12/2015